+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## future
+### New features
+- Add codes `ERASE_BUFFER` and `CLEAR`.
+### Fixes
++ Fix `ERASE_ALL`.
+
 ## v2.0.0
 - New methods for `Terminal`: `has_input`, `wait_for_input`,
   `read_ambiguous_timeout`, `read_timeout`, `read_raw`, `read_raw_timeout`,